modules/grpc-location-client/requirements.txt
kafka-python
# keyboard
grpcio-tools